+++ marshmallow/schema.py
@@ -213,6 +213,8 @@ class SchemaOpts(object):
         self.ordered = getattr(meta, 'ordered', False)
         self.index_errors = getattr(meta, 'index_errors', True)
         self.include = getattr(meta, 'include', {})
+        self.load_only = getattr(meta, 'load_only', ())
+        self.dump_only = getattr(meta, 'dump_only', ())
 
 
 class BaseSchema(base.SchemaABC):
@@ -259,6 +261,9 @@ class BaseSchema(base.SchemaABC):
         serialized results if ``value`` is `None`.
     :param dict context: Optional context passed to :class:`fields.Method` and
         :class:`fields.Function` fields.
+    :param tuple load_only: A list or tuple of fields to skip during serialization
+    :param tuple dump_only: A list or tuple of fields to skip during
+        deserialization, read-only fields
     """
     TYPE_MAPPING = {
         text_type: fields.String,
@@ -325,6 +330,8 @@ class BaseSchema(base.SchemaABC):
             `collections.OrderedDict`.
         - ``index_errors``: If `True`, errors dictionaries will include the index
             of invalid items in a collection.
+        - ``load_only``: Tuple or list of fields to exclude from serialized results.
+        - ``dump_only``: Tuple or list of fields to exclude from deserialization
 
         .. versionchanged:: 2.0.0
             `__preprocessors__` and `__data_handlers__` are deprecated. Use
@@ -333,7 +340,7 @@ class BaseSchema(base.SchemaABC):
         pass
 
     def __init__(self, extra=None, only=(), exclude=(), prefix='', strict=False,
-                 many=False, context=None):
+                 many=False, context=None, load_only=(), dump_only=()):
         # copy declared fields from metaclass
         self.declared_fields = copy.deepcopy(self._declared_fields)
         self.many = many
@@ -342,6 +349,8 @@ class BaseSchema(base.SchemaABC):
         self.prefix = prefix
         self.strict = strict or self.opts.strict
         self.ordered = self.opts.ordered
+        self.load_only = set(load_only) or set(self.opts.load_only)
+        self.dump_only = set(dump_only) or set(self.opts.dump_only)
         #: Dictionary mapping field_names -> :class:`Field` objects
         self.fields = self.dict_class()
         #: Callable marshalling object
@@ -589,9 +598,15 @@ class BaseSchema(base.SchemaABC):
         """Override-able method that defines how to create the final deserialization
         output. Defaults to noop (i.e. just return ``data`` as is).
 
+        .. note::
+
+            This method will only be invoked if when the input data are completely valid.
+
         :param dict data: The deserialized data.
 
         .. versionadded:: 1.0.0
+        .. versionchanged:: 2.0.0
+            Only invoked when data are valid.
         """
         return data
 
@@ -623,7 +638,6 @@ class BaseSchema(base.SchemaABC):
         else:
             preprocessors = []
 
-        postprocess_funcs = [self.make_object] if postprocess else []
         result = self._unmarshal(
             data,
             self.fields,
@@ -631,7 +645,6 @@ class BaseSchema(base.SchemaABC):
             strict=self.strict,
             validators=validators,
             preprocess=preprocessors,
-            postprocess=postprocess_funcs,
             dict_class=self.dict_class,
             index_errors=self.opts.index_errors,
         )
@@ -644,6 +657,12 @@ class BaseSchema(base.SchemaABC):
             self.__error_handler__(errors, data)
 
         result = self._invoke_load_processors(POST_LOAD, result, many)
+
+        if not errors and postprocess:
+            if many:
+                result = [self.make_object(each) for each in result]
+            else:
+                result = self.make_object(result)
 
         return result, errors
 
@@ -674,18 +693,25 @@ class BaseSchema(base.SchemaABC):
 
     def __set_field_attrs(self, fields_dict):
         """Update fields with values from schema.
+
+        Also set field load_only and dump_only values if field_name was
+        specified in ``class Meta``.
         """
         for field_name, field_obj in iteritems(fields_dict):
             try:
+                if field_name in self.load_only:
+                    field_obj.load_only = True
+                if field_name in self.dump_only:
+                    field_obj.dump_only = True
                 field_obj._add_to_schema(field_name, self)
             except TypeError:
                 # field declared as a class, not an instance
                 if (isinstance(field_obj, type) and
                         issubclass(field_obj, base.FieldABC)):
                     msg = ('Field for "{0}" must be declared as a '
-                                    'Field instance, not a class. '
-                                    'Did you mean "fields.{1}()"?'
-                                    .format(field_name, field_obj.__name__))
+                           'Field instance, not a class. '
+                           'Did you mean "fields.{1}()"?'
+                           .format(field_name, field_obj.__name__))
                     raise TypeError(msg)
         return fields_dict